==> jupyter/4_compranet_dedup_public.ipynb <==
# Deduplicated public buyers

import pandas as pd
import numpy as np

CNTS = '/home/rdora/declaranet/data/pre-process/contratos_3.csv'
cnts = pd.read_csv(CNTS)

public = cnts.groupby(['buyer', 'buyer_state', 'ramo', 'gvmnt_level'])[
    ['buyer', 'buyer_state', 'ramo', 'gvmnt_level']].head()

PUBLIC = '/home/rdora/declaranet/data/dedup/public.csv'
public.to_csv(PUBLIC, index=False)

public.buyer_state.unique()

# Run Dedup

Here, we run the script `~/declaranet/python/deduplicate.py` to create a cluster for each record.

PUBLIC_2 = '/home/rdora/declaranet/data/dedup/output_public.csv'
public_cl = pd.read_csv(PUBLIC_2)

# Groupby

def get_mode(group):
    vc = group.value_counts()
    if vc.shape[0]:
        return vc.index[0]
    else:
        return np.nan

gb_public = public_cl.groupby('Cluster ID')['buyer', 'buyer_state'].agg(get_mode).reset_index()

## Companies

public = cnts.copy()

public = pd.merge(public,
                public_cl[['Cluster ID', 'buyer', 'buyer_state']],
                on=['buyer', 'buyer_state'],
                how='left')

public = public.drop(['buyer', 'buyer_state'], axis=1)

public = pd.merge(public, gb_public, how='left')

print("New contracts with Buyer State:",
      cnts.buyer_state.isna().sum() - public.buyer_state.isna().sum())

print("Number of duplicated UCs:",
     cnts.buyer.nunique() - public.buyer.nunique())

buyer = set(cnts.buyer.unique())
buyer_2 = set(public.buyer.unique())
miss = buyer - buyer_2

print(f"Number of deduplicated edges: {cnts[cnts.buyer.isin(miss)].shape[0]:,}")
