==> jupyter/5_compranet_adjacency_list.ipynb <==
# Create network adjacency list

import pandas as pd
import numpy as np

CNTS = '/home/rdora/declaranet/data/pre-process/contratos_3.csv'
cnts = pd.read_csv(CNTS, parse_dates=['start_date'])

cnts = cnts[cnts.supplier_country=='MX']

cnts = cnts[~cnts.supplier_state.isin(['BB-05', 'US-KS',
       'US-CA', 'ES-PO', 'US-WA'])]

cnts = cnts.dropna(subset=['supplier_state', 'buyer_state'])

drop_cols = [
    'prop_window',
    'gvmnt_level',
    'contract_type',
    'id',
    'supplier_country',
    'desc_ramo',
    'ramo']

cnts_local = cnts[cnts.gvmnt_level != 'F']

cnts_federal = cnts[cnts.gvmnt_level == 'F']

cnts = cnts.drop(drop_cols, axis=1)

cnts_local = cnts_local.drop(drop_cols, axis=1)

cnts_federal = cnts_federal.drop(drop_cols, axis=1)

cnts = cnts.dropna()

cnts_local = cnts_local.dropna()

cnts_federal = cnts_federal.dropna()

# Save as csv

buyers = cnts.buyer.unique()
suppliers = cnts.supplier.unique()

dbuyers = dict(zip(buyers, [f'b{i}' for i in range(len(buyers))]))
dsuppliers = dict(zip(suppliers, [f's{i}' for i in range(len(suppliers))]))

cnts['buyer'] = cnts['buyer'].apply(lambda x: dbuyers[x])

cnts['supplier'] = cnts['supplier'].apply(lambda x: dsuppliers[x])

cnts_local['buyer'] = cnts_local['buyer'].apply(lambda x: dbuyers[x])

cnts_local['supplier'] = cnts_local['supplier'].apply(lambda x: dsuppliers[x])

cnts_federal['buyer'] = cnts_federal['buyer'].apply(lambda x: dbuyers[x])

cnts_federal['supplier'] = cnts_federal['supplier'].apply(lambda x: dsuppliers[x])

OUT = '/home/rdora/declaranet/data/network/edge_list.csv'
cnts.to_csv(OUT, index=False)

OUT_LOCAL = '/home/rdora/declaranet/data/network/edge_list_local.csv'
cnts_local.to_csv(OUT_LOCAL, index=False)

OUT_FEDERAL = '/home/rdora/declaranet/data/network/edge_list_federal.csv'
cnts_federal.to_csv(OUT_FEDERAL, index=False)
