==> jupyter/11_CRI_Stats.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns; sns.set(style='white')
import matplotlib.pyplot as plt

cnts = pd.read_csv("../data/pre-process/contratos_4.csv")

covariates = ['code_b',
              'same_state',
              'person',
              'single_bid',
              'daily_price',
              'amount',
              'file_size',
              'atom',
              'CRI']

table = cnts[covariates]

table = table.groupby('code_b').mean()

table = pd.merge(table,
                 cnts.groupby('code_b').size().rename('Contracts'),
                 left_on="code_b",
                 right_index=True,
                 how="left")

order = [
         'Contracts',
         'atom',
         'same_state',
         'person',
         'single_bid',
         'amount',
         'daily_price',
         'file_size',
         'CRI']

print(table[order].round(2).to_latex())
