==> jupyter/12_cores.ipynb <==
import pandas as pd
import networkx as nx
import numpy as np
import seaborn as sns; sns.set(style='white')
import matplotlib.pyplot as plt

from unidecode import unidecode

cnts = pd.read_csv("../data/pre-process/contratos_4.csv",
                   parse_dates=['start_date'])

## One last dedup

By removing accents and so forth

def normalize(x):
    return unidecode(x.strip().lower())

cnts.supplier.nunique()

cnts.buyer.nunique()

cnts['buyer'] = cnts.buyer.apply(normalize)

cnts['supplier'] = cnts.supplier.apply(normalize)

cnts.supplier.nunique()

cnts.buyer.nunique()

cnts['year'] = cnts.start_date.dt.year

cnts.groupby('code_b').single_bid.mean()

net = cnts.groupby(['code_b', 'code_s', 'buyer', 'supplier', 'year']).agg(
    weight=('amount', 'sum'),
    cri = ('CRI', 'mean')).reset_index()

years = net.year.unique()

Gs = []
for year in sorted(years):
    Gs.append(nx.from_pandas_edgelist(
        net[net.year==year],
        source='buyer',
        target='supplier',
        edge_attr='cri'))

nx.con

data = []
for i, G in enumerate(Gs):
    year = 2011 + i
    nodes = len(G)
    edges = len(G.edges)
    density = nx.density(G)
    largest_component = len(max(nx.connected_components(G), key=len))
    number_components = nx.number_connected_components(G)
    degrees = list(dict(G.degree).values())
    mean_degree = np.mean(degrees)
    max_degree = np.max(degrees)
    min_degree = np.min(degrees)
    std_degree = np.std(degrees)
    data.append([year, nodes, edges, density, largest_component, number_components, mean_degree,
                max_degree, min_degree, std_degree])
df = pd.DataFrame(data, columns=['year', 'nodes', 'edges', 'density', 'largest_component',
                           'number_components', 'mean_degree', 'max_degree', 'min_degree',
                           'std_degree'])

df

fig, axs = plt.subplots(3,3, figsize=(12, 9), sharex=True, sharey=True)
for i, G in enumerate(Gs):
    x, y = divmod(i, 3)
    sns.distplot(list(nx.core_number(G).values()),
            kde=False, ax=axs[x][y], bins=20)
    axs[x][y].set_title(f'Year {2011 + i}')
    axs[x][y].set_yscale('log')

fig, axs = plt.subplots(3,3, figsize=(12, 9), sharex=True, sharey=True)
for i, G in enumerate(Gs):
    x, y = divmod(i, 3)
    sns.distplot(list(dict(G.degree()).values()),
            kde=False, ax=axs[x][y], bins=20)
    axs[x][y].set_title(f'Year {2011 + i}')
    axs[x][y].set_yscale('log')
